Omaha #5515 - ensure a supposedly unused message is never removed from cache within the current purge cycle.
Handles the case (encountered at OAX) in which a message may be loaded into cache, but the purge
task runs before the associated playlist can be correctly classified. This changeset adds a grace
period for the case that a message is due to be removed from cache because it is supposedly no
longer referenced by any current playlists. This has no effect on cache removal due to expiration.